Simplify agent sync package
