--- .github/scripts/agent_sync/exceptions.py
@@ -1,5 +1,0 @@
-__all__ = ["AgentSyncConfigError"]
-
-
-class AgentSyncConfigError(ValueError):
-    """Report invalid canonical agent configuration."""
