+++ .github/scripts/agent_sync/generation.py
@@ -1,21 +1,27 @@
 import json
+import tomllib
 
-from agent_sync.codex import generate_codex_config
+from agent_sync.configuration import AgentSyncConfigError
+from agent_sync.instructions import (
+    assemble_universal_agent_instructions,
+    generate_rule_outputs,
+    generate_skill_outputs,
+)
 from agent_sync.markdown import ensure_trailing_newline
+from agent_sync.models.configuration import (
+    AgentModelOverride,
+    AgentSyncSettings,
+    CodexSettings,
+    Provider,
+    Workspace,
+)
 from agent_sync.models.outputs import OutputFile, OutputKind
-from agent_sync.models.providers.codex import CodexSettings
-from agent_sync.models.providers.providers import Provider
-from agent_sync.models.settings import AgentModelOverride, AgentSyncSettings
-from agent_sync.models.workspace import Workspace
 from agent_sync.providers import (
     generate_agent_outputs,
     generate_command_outputs,
     generate_hook_outputs,
     generate_settings_outputs,
 )
-from agent_sync.rules import assemble_universal_agent_instructions, generate_rule_outputs
-from agent_sync.skills import generate_skill_outputs
-from agent_sync.validation import run_validations
 
 __all__ = ["generate_outputs"]
 
@@ -26,8 +32,6 @@ def generate_outputs(
     model_overrides: dict[str, AgentModelOverride],
 ) -> list[OutputFile]:
     """Generate every provider output from canonical agent sources."""
-
-    run_validations(workspace, settings)
 
     return [
         *generate_skill_outputs(workspace),
@@ -105,3 +109,27 @@ def render_codex_settings(settings: CodexSettings) -> str:
     """Serialize canonical Codex settings with the synchronized instruction capacity."""
 
     return ensure_trailing_newline(json.dumps(settings.model_dump(exclude_none=True), indent=2))
+
+
+def generate_codex_config(settings: CodexSettings) -> str:
+    """Render Codex configuration entirely from canonical settings."""
+
+    lines = ["# >>> agent-sync managed Codex settings >>>"]
+
+    if settings.model is not None:
+        lines.append(f"model = {json.dumps(settings.model, ensure_ascii=False)}")
+
+    lines.extend(
+        (
+            f"project_doc_max_bytes = {settings.project_doc_max_bytes}",
+            "# <<< agent-sync managed Codex settings <<<",
+        )
+    )
+    rendered = ensure_trailing_newline("\n".join(lines))
+
+    try:
+        tomllib.loads(rendered)
+    except tomllib.TOMLDecodeError as exc:
+        raise AgentSyncConfigError(f"Generated .codex/config.toml is invalid TOML: {exc}") from exc
+
+    return rendered
