+++ .github/scripts/agent_sync/instructions.py
@@ -1,24 +1,101 @@
+import logging
 from pathlib import Path
 from typing import Final
 
+from agent_sync.configuration import AgentSyncConfigError
 from agent_sync.markdown import (
     ensure_trailing_newline,
     normalize_rule_source,
     parse_markdown_file,
     validate_slug,
 )
-from agent_sync.models.frontmatter import RuleFrontMatter
+from agent_sync.models.configuration import Provider, Workspace
+from agent_sync.models.documents import RuleFrontMatter, SkillFrontMatter
 from agent_sync.models.outputs import OutputFile, OutputKind
-from agent_sync.models.providers.providers import Provider
-from agent_sync.models.workspace import Workspace
 
 __all__ = [
     "assemble_universal_agent_instructions",
     "generate_rule_outputs",
-    "is_managed_codex_rule",
+    "generate_skill_outputs",
 ]
 
+logger = logging.getLogger(__name__)
+
 CODEX_RULE_MARKER: Final[str] = "# Generated by agent_sync from .agents/rules/"
+
+
+def generate_skill_outputs(workspace: Workspace) -> list[OutputFile]:
+    """Generate provider skill links from canonical skill directories."""
+
+    skills_dir = workspace.agents / "skills"
+
+    if not skills_dir.exists():
+        return []
+
+    outputs: list[OutputFile] = []
+    for skill_dir in sorted(path for path in skills_dir.iterdir() if path.is_dir()):
+        slug = validate_slug(skill_dir.name, skill_dir)
+        source_path = skill_dir / "SKILL.md"
+
+        if not source_path.exists():
+            logger.warning("Missing SKILL.md in %s", skill_dir)
+            continue
+
+        front_matter, _ = parse_markdown_file(source_path, SkillFrontMatter)
+        validate_canonical_skill_metadata(front_matter, source_path)
+        outputs.extend(
+            OutputFile(
+                target_path=workspace.root / f".{provider.value}" / "skills" / slug,
+                content="",
+                kind=skill_link_kind(provider),
+                slug=slug,
+                source_path=source_path,
+                link_target=skill_dir,
+            )
+            for provider in (Provider.CLAUDE, Provider.CURSOR)
+        )
+        outputs.append(
+            OutputFile(
+                target_path=workspace.root / ".codex" / "skills" / slug,
+                content="",
+                kind=OutputKind.CODEX_SKILL,
+                slug=slug,
+                source_path=source_path,
+                link_target=skill_dir,
+            )
+        )
+
+    return outputs
+
+
+def validate_canonical_skill_metadata(
+    front_matter: SkillFrontMatter,
+    source_path: Path,
+) -> None:
+    """Require metadata shared by every canonical skill."""
+
+    if not front_matter.name or not front_matter.description:
+        raise AgentSyncConfigError(
+            f"Skill {source_path} must define non-empty name and description front matter"
+        )
+
+    if front_matter.name != source_path.parent.name:
+        raise AgentSyncConfigError(
+            f"Skill {source_path} must use directory name {source_path.parent.name!r} "
+            f"as its front matter name, not {front_matter.name!r}"
+        )
+
+
+def skill_link_kind(provider: Provider) -> OutputKind:
+    """Return the directory-link output kind for one provider."""
+
+    match provider:
+        case Provider.CLAUDE:
+            return OutputKind.CLAUDE_SKILL
+        case Provider.CURSOR:
+            return OutputKind.CURSOR_SKILL
+        case Provider.CODEX:
+            raise ValueError("Codex skills use their dedicated directory-link output")
 
 
 def generate_rule_outputs(workspace: Workspace) -> list[OutputFile]:
@@ -116,10 +193,7 @@ def render_rule_section(filename: str, front_matter: RuleFrontMatter, body: str)
 def render_rule_scope(front_matter: RuleFrontMatter) -> str:
     """Render the universal applicability note for one rule."""
 
-    if isinstance(front_matter.globs, str):
-        patterns = [front_matter.globs]
-    else:
-        patterns = front_matter.globs or []
+    patterns = [front_matter.globs] if isinstance(front_matter.globs, str) else front_matter.globs or []
 
     if patterns:
         rendered_patterns = ", ".join(f"`{pattern}`" for pattern in patterns)
@@ -129,9 +203,3 @@ def render_rule_scope(front_matter: RuleFrontMatter) -> str:
         return "> Apply this rule only when it is explicitly relevant to the current task."
 
     return ""
-
-
-def is_managed_codex_rule(content: str) -> bool:
-    """Return whether a Starlark policy was generated by agent sync."""
-
-    return content.startswith(CODEX_RULE_MARKER)
