--- .github/scripts/agent_sync/models/providers/claude.py
@@ -1,7 +1,0 @@
-from agent_sync.models.providers.providers import ProviderSettings
-
-__all__ = ["ClaudeSettings"]
-
-
-class ClaudeSettings(ProviderSettings):
-    """Validate canonical Claude project settings."""
