--- .github/scripts/agent_sync/models/providers/cursor.py
@@ -1,7 +1,0 @@
-from agent_sync.models.providers.providers import ProviderSettings
-
-__all__ = ["CursorSettings"]
-
-
-class CursorSettings(ProviderSettings):
-    """Validate canonical Cursor project settings."""
