+++ .github/scripts/agent_sync/providers.py
@@ -8,11 +8,14 @@ from agent_sync.markdown import (
     render_front_matter,
     validate_slug,
 )
-from agent_sync.models.frontmatter import AgentFrontMatter, CommandFrontMatter
+from agent_sync.models.configuration import (
+    AgentModelOverride,
+    AgentSyncSettings,
+    Provider,
+    Workspace,
+)
+from agent_sync.models.documents import AgentFrontMatter, CommandFrontMatter
 from agent_sync.models.outputs import OutputFile, OutputKind
-from agent_sync.models.providers.providers import Provider
-from agent_sync.models.settings import AgentModelOverride, AgentSyncSettings
-from agent_sync.models.workspace import Workspace
 from agent_sync.storage import read_text
 
 __all__ = [
